refactor : package 분리
view package와 domain package를 만들어서 클래스를 분리합니다.
