+++ src/main/java/ladder/Main.java
@@ -1,0 +1,10 @@
+package ladder;
+
+import ladder.domain.LadderGame;
+
+public class Main {
+    public static void main(String[] args) {
+        LadderGame ladderGame = new LadderGame();
+        ladderGame.start();
+    }
+}
